use a single highlight group `CursorWord` in favor of Vim 7.4.682
Historically this plugin used two highlight groups `CursorWord0` and
`CursorWord1`. This was required when I initially created the plugin, in order
to use different highlight groups in the current line and other places. It was
important to respect the background colors of `CursorLine` when `'cursorline'`
is on. However, Vim 7.4.682 merges the colors of matching highlight group and
that of `CursorLine`, so we don't need to use two highlight groups anymore.
Also, using the legacy `redir` command caused various problems.

This commit changes the plugin to depend on only a single highlight group
`CursorWord`. This change breaks the behavior for users customizing the
highlight group colors but this is intended changes to encourage the users
removing unnecessary highlight group configuration. Furthermore, this
plugin starts to specify the `default` attribute on setting the highlight
group to avoid overriding customized configuration. This change could
eliminate unnecessary auto-commands and simplify customizations.